Added enemies with state machine
Added enemies with state machine and a parent object that will easily be inheritable across any other enemy in the game. For now the enemies behave Idle > Chase > Alert > Idle

Starting as idle, chasing the player as soon as spotted, being in alert phase for a bit after breaking contact, returning to its original position in the idle state.

The next step should be to make the attack damage the enemies.