Many changens, I forgot to commit all of them
So now we have art for the player, tilesets and enemies
I changed a lot of how the code works, instead of variable based, I made it animation based.

There were a lot of bugs I had to fix but its working semi intended now. Now I need to... do final touches, balancing, implement sounds and menu, and we're good to launch it.